i added everything this time
I made too much progress to remember what I changed. just look at the
code :c Got rid of dynamic js and modernizr because I don't think it was
working and I don't care enough about it to fix it. Finally got gallery
images to center in the accordion. Finished writing most of the content,
but it's subject to change. made a new version of the candle gif to
spruce up the home page. Added more break points